+++ R/quantityreport-class.R
@@ -55,6 +55,7 @@ as_quantityreport <- function(x, .show_check = TRUE) {
   x %>%
     add_missing_spec_tables(table_spec, check) %>%
     add_missing_spec_cols(table_spec, new_name = "field") %>%
+    drop_na_optional_spec_tables(table_spec) %>%
     new_quantityreport()
 }