+++ data-raw/quantity_spec.R
@@ -8,7 +8,7 @@ quantity_tables <- readxl::read_excel("data-raw/quantity-tables.xlsx",
 
 quantity_fields <- readxl::read_excel("data-raw/quantity-tables.xlsx",
                                       sheet = "fields",
-                                      col_types = c(rep("text", 5), "logical", rep("text", 2))) %>%
+                                      col_types = c(rep("text", 5), "logical", rep("text", 2), "logical")) %>%
   dplyr::mutate(safe_name = dplyr::coalesce(safe_name, snake_name))
 
 quantity_spec <- list(tables = quantity_tables,